Resolving struct member type inconsistency, which could lead to segfaults

+++ inst/include/niftilib/nifti2_io.h
@@ -234,7 +234,7 @@ typedef struct {
   int nv ;                      /*!< dimensions of grid array             */
   int nw ;                      /*!< dimensions of grid array             */
   int dim[8] ;                  /*!< dim[0]=ndim, dim[1]=nx, etc.         */
-  int64_t nvox ;                /*!< number of voxels = nx*ny*nz*...*nw   */
+  size_t nvox ;                 /*!< number of voxels = nx*ny*nz*...*nw   */
   int nbyper ;                  /*!< bytes per voxel, matches datatype    */
   int datatype ;                /*!< type of data in voxels: DT_* code    */
